+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.4
+* solidTwo added for border layout
+* solidTwo color groups can access it like primarySolidTwo(), successSolidTwo() & so on
 ## 0.0.3
 * Major part cover Text, Button, Icon, Card, Container, Column, Row, Center, Stack, Positioned, Wrap, GestureDetector
 * You can also use with less properties Image, ListTile, RadioListTile,